Remove AXTLS from core
axTLS has been deprecated for over a year now, so finally remove it from
the code.

Rename BearSSL* examples to plain WiFiSecure*
