+++ libraries/ESP8266WiFi/examples/WiFiServerSecureClientCert/WiFiServerSecureClientCert.ino
